+++ opentelemetry/src/context/mod.rs
@@ -1,4 +1,6 @@
 mod context_store;
+mod future_ext;
 
 pub use context_store::Context;
 pub use context_store::ContextGuard;
+pub use future_ext::FutureExt;
